Update master to 0.29.1

+++ lib/AL_USDMaya/AL/usdmaya/nodes/RendererManager.cpp
@@ -214,7 +214,7 @@ bool RendererManager::setInternalValueInContext(const MPlug& plug, const MDataHa
   if (plug == m_rendererPlugin)
   {
     short index = dataHandle.asShort();
-    if (index >= 0 && index < m_rendererPluginsNames.length())
+    if (index >= 0 && index < short(m_rendererPluginsNames.length()))
     {
       MPlug plug(thisMObject(), m_rendererPluginName);
       plug.setString(m_rendererPluginsNames[index]);
@@ -272,7 +272,7 @@ void RendererManager::changeRendererPlugin(ProxyShape* proxy, bool creation)
       if (rendererId == 0 && creation)
         return;
       
-      assert(rendererId < m_rendererPluginsTokens.size());
+      assert(static_cast<size_t>(rendererId) < m_rendererPluginsTokens.size());
       TfToken plugin = m_rendererPluginsTokens[rendererId];
       if (!proxy->engine()->SetRendererPlugin(plugin))
       {
